refactor(admin): split dashboard section modules
Why:
- keep the dashboard route imports stable while separating site copy, health, business, and live mini sections
- reduce the mixed dashboard-sections module into focused server component boundaries

Verify:
- pnpm exec prettier --check apps/admin/src/app/dashboard-sections.tsx apps/admin/src/app/dashboard-site-copy-section.tsx apps/admin/src/app/dashboard-health-section.tsx apps/admin/src/app/dashboard-business-sections.tsx apps/admin/src/app/dashboard-live-section.tsx
- git diff --check -- apps/admin/src/app/dashboard-sections.tsx apps/admin/src/app/dashboard-site-copy-section.tsx apps/admin/src/app/dashboard-health-section.tsx apps/admin/src/app/dashboard-business-sections.tsx apps/admin/src/app/dashboard-live-section.tsx
- pnpm --filter @anipotts/admin typecheck
- pnpm --filter @anipotts/admin lint
- pnpm --filter @anipotts/admin build
- pnpm validate
